requirements.txt
Pillow==10.0.1
fpdf2==2.7.6
reportlab==4.0.4 